feat(ui): organize migrations files
