+++ invokeai/frontend/web/src/features/nodes/types/v1/fieldTypeMap.ts
@@ -1,4 +1,4 @@
-import { FieldType, StatefulFieldType } from '../../field';
+import { FieldType, StatefulFieldType } from '../field';
 import { FieldTypeV1 } from './workflowV1';
 
 /**
